+++ tests/__init__.py
@@ -1,13 +1,7 @@
 # -*- coding:utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import logging
-
 import six
-
-from hfut import logger
-
-logger.setLevel(logging.DEBUG)
 
 
 class TestBase(object):